+++ src/tables/const_column.cpp
@@ -3,13 +3,13 @@
 namespace tables
 {
 
-    const_column::const_column(const boost::any& value, uint64_t rows)
+    const_column::const_column(const boost::any& value, size_t rows)
         : m_value(value), m_rows(rows)
     {
         assert(!value.empty());
     }
 
-    boost::any const_column::value(uint64_t row_index) const
+    boost::any const_column::value(size_t row_index) const
     {
         // did you forget to add rows?
         assert(m_rows > 0);
@@ -30,7 +30,7 @@ namespace tables
         return boost::any();
     }
 
-    void const_column::add_rows(uint64_t rows)
+    void const_column::add_rows(size_t rows)
     {
         m_rows += rows;
     }
@@ -40,12 +40,12 @@ namespace tables
         m_rows += 1;
     }
 
-    uint64_t const_column::rows() const
+    size_t const_column::rows() const
     {
         return m_rows;
     }
 
-    uint64_t const_column::empty_rows() const
+    size_t const_column::empty_rows() const
     {
         return 0;
     }
